+++ src/person/person.controller.ts
@@ -1,28 +1,56 @@
 import { Response, Request } from 'express'
 import { hash } from 'bcrypt-ts'
-import Person, { IPerson } from './person.model.js'
+import Person from './person.model'
 import { MongoServerError } from 'mongodb'
+import { Result, validationResult } from 'express-validator'
+import { PaginateOptions } from 'mongoose'
+import Competitor from '../competitor/competitor.model'
 
 export async function findAll(req: Request, res: Response) {
+  const result: Result = validationResult(req)
+  const errors = result.array()
+
+  if (!result.isEmpty()) {
+    return res.status(400).json({ errors: errors })
+  }
+
   var filter = req.query.filter
+  var page = Number(req.query.page)
+  const options: PaginateOptions = {
+    page: page,
+    limit: 10,
+    select: '-password',
+    sort: { updatedAt: -1 },
+  }
 
   if (filter) {
     filter = filter.toString()
-    const persons = await Person.find({
-      $or: [
-        { dni: { $regex: new RegExp(filter, 'i') } },
-        { firstName: { $regex: new RegExp(filter, 'i') } },
-        { lastName: { $regex: new RegExp(filter, 'i') } },
-      ],
-    })
+    const persons = await Person.paginate(
+      {
+        $or: [
+          { dni: { $regex: new RegExp(filter, 'i') } },
+          { firstName: { $regex: new RegExp(filter, 'i') } },
+          { lastName: { $regex: new RegExp(filter, 'i') } },
+        ],
+      },
+      options
+    )
     return res.json(persons)
   }
 
-  const persons = await Person.find().select('-password')
+  const persons = await Person.paginate({}, options)
+
   res.json(persons)
 }
 
 export async function findOne(req: Request, res: Response) {
+  const result: Result = validationResult(req)
+  const errors = result.array()
+
+  if (!result.isEmpty()) {
+    return res.status(400).json({ errors: errors })
+  }
+
   try {
     const person = await Person.findById(req.params.id).select('-password')
 
@@ -42,29 +70,31 @@ export async function findOne(req: Request, res: Response) {
 }
 
 export async function add(req: Request, res: Response) {
+  const result: Result = validationResult(req)
+  const errors = result.array()
+
+  if (!result.isEmpty()) {
+    return res.status(400).json({ errors: errors })
+  }
+
   const personInput = new Person(req.body)
 
-  if (personInput.password) {
-    const hashedPassword = await hash(personInput.password, 10)
-    personInput.password = hashedPassword
-  }
-
-  if (personInput.birthdate && new Date(personInput.birthdate).toString() === 'Invalid Id') {
-    return res.status(400).send({ message: 'La fecha de nacimineto ingresada no es valida' })
-  }
+  const hashedPassword = await hash(personInput.password, 10)
+  personInput.password = hashedPassword
 
   try {
     const person = await personInput.save()
     const personObject = person.toObject()
     const { password, ...personWPassword } = personObject
-    return res.status(201).json({ message: 'Person created', data: personWPassword })
+
+    return res.status(201).json({ message: 'Persona creada', data: personWPassword })
   } catch (err) {
     const mongoErr: MongoServerError = err as MongoServerError
     if (mongoErr.code === 11000) {
       if (mongoErr.keyPattern && mongoErr.keyPattern.email === 1) {
-        return res.status(400).send({ message: 'Email Repetido' })
+        return res.status(409).send({ message: 'Email Repetido' })
       } else {
-        return res.status(400).send({ message: 'DNI Repetido' })
+        return res.status(409).send({ message: 'DNI Repetido' })
       }
     } else if (mongoErr.name === 'ValidationError') {
       return res.status(400).send({ message: 'Falta un atributo requerido' })
@@ -77,15 +107,23 @@ export async function add(req: Request, res: Response) {
 
 export async function update(req: Request, res: Response) {
   const id = req.params.id
+  const result: Result = validationResult(req)
+  const errors = result.array()
+
+  if (!result.isEmpty()) {
+    return res.status(400).json({ errors: errors })
+  }
+
   try {
+    if (req.body.password) {
+      const hashedPassword = await hash(req.body.password, 10)
+      req.body.password = hashedPassword
+    }
+
     const person = await Person.findByIdAndUpdate(id, req.body, { new: true }).select('-password')
 
     if (!person) {
       return res.status(404).send({ message: 'Persona no encontrada' })
-    }
-
-    if (req.body.birthdate && new Date(req.body.birthdate).toString() === 'Invalid Id') {
-      return res.status(400).send({ message: 'La fecha de nacimineto ingresada no es valida' })
     }
 
     return res.status(200).json(person)
@@ -93,9 +131,9 @@ export async function update(req: Request, res: Response) {
     const mongoErr: MongoServerError = err as MongoServerError
     if (mongoErr.code === 11000) {
       if (mongoErr.keyPattern && mongoErr.keyPattern.email === 1) {
-        return res.status(400).send({ message: 'Email Repetido' })
+        return res.status(409).send({ message: 'Email Repetido' })
       } else {
-        return res.status(400).send({ message: 'DNI Repetido' })
+        return res.status(409).send({ message: 'DNI Repetido' })
       }
     } else {
       if (err instanceof Error && err.name === 'CastError') {
@@ -111,13 +149,28 @@ export async function update(req: Request, res: Response) {
 export async function remove(req: Request, res: Response) {
   const id = req.params.id
 
+  //errores Bad Request
+  const result: Result = validationResult(req)
+  const errors = result.array()
+
+  if (!result.isEmpty()) {
+    return res.status(400).json({ errors: errors })
+  }
+
   try {
+    const competitors = await Competitor.find({ person: id })
+
+    if (competitors.length !== 0) {
+      return res.status(409).send({ message: 'El socio tiene inscripciones cargadas' })
+    }
+
     const person = await Person.findByIdAndDelete(id)
 
     if (!person) {
       return res.status(404).send({ message: 'Persona no encontrada' })
     }
-    return res.status(200).send({ message: 'Person deleted', data: person })
+
+    return res.status(200).send({ message: 'Persona eliminada', data: person })
   } catch (err) {
     if (err instanceof Error && err.name === 'CastError') {
       console.log(err)
